% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryCalibrationMetrics.R
\name{binaryCalibrationMetrics}
\alias{binaryCalibrationMetrics}
\title{Calibration Metrics}
\usage{
binaryCalibrationMetrics(model, testing_dataset, bin_cnt = 10)
}
\arguments{
\item{model}{Model glm fit.}

\item{testing_dataset}{The dataset to test the model.}

\item{bin_cnt}{The number of bins to calibrate the probabilities.}
}
\value{
A plot.
}
\description{
Calculates the mean predicted value and proportion of positives for a model.
}
\keyword{binary}
\keyword{plot,}
\keyword{precision,}
\keyword{recall,}
