% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassificationMetrics.R
\name{binaryClassificationMetrics}
\alias{binaryClassificationMetrics}
\title{Binary Classification Metrics}
\usage{
binaryClassificationMetrics(
  model,
  testing_dataset,
  selected_metric = c("ACC", "ERR", "FPR", "TPR", "FNR", "TNR", "PPV", "NPV", "F",
    "RPP", "RNP")
)
}
\arguments{
\item{model}{Model glm fit.}

\item{testing_dataset}{The dataset to test the model.}

\item{selected_metric}{The selected metric(s) to pull.}
}
\value{
A vector of metrics.
}
\description{
This function calculates classification evaluation metrics for binary outcomes.
}
\keyword{binary,}
\keyword{classification}
\keyword{metric,}
\keyword{model,}
