% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryModelMetrics.R
\name{binaryModelMetrics}
\alias{binaryModelMetrics}
\title{Binary Model Metrics}
\usage{
binaryModelMetrics(
  model,
  testing_dataset,
  selected_metric = c("BRI", "MSE", "RMSE", "ROC", "LLS", "AIC", "BIC", "PRC", "LOG",
    "LLM", "LL0", "CXS", "NAG", "DEV", "SMD", "TJU", "MCF", "PEA", "HOS")
)
}
\arguments{
\item{model}{Model glm fit.}

\item{testing_dataset}{The dataset to test the model.}

\item{selected_metric}{The selected metric(s).}
}
\value{
A vector of metrics.
}
\description{
This function calculates model evaluation metrics for binary outcomes.
}
\keyword{binary}
\keyword{metric,}
\keyword{model,}
