% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationPlot.R
\name{calibrationPlot}
\alias{calibrationPlot}
\title{Calibration Plot}
\usage{
calibrationPlot(mev, pop)
}
\arguments{
\item{mev}{Mean estimated value for each bin.}

\item{pop}{Proportion of positives.}
}
\value{
A calibration plot.
}
\description{
Plots the calibration curve.
}
\keyword{binary,}
\keyword{calibration,}
\keyword{estimated}
\keyword{mean}
\keyword{of}
\keyword{plot,}
\keyword{positives}
\keyword{proportion}
\keyword{value,}
