% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcPlot.R
\name{prPlot}
\alias{prPlot}
\title{Precision Recall Plot}
\usage{
prPlot(ppv, tpr)
}
\arguments{
\item{ppv}{Positive predicted value rate. Also known as precision.}

\item{tpr}{True positive rate. Also known as recall.}
}
\value{
A plot.
}
\description{
Plots the precision recall curve.
}
\keyword{binary}
\keyword{plot,}
\keyword{precision,}
\keyword{recall,}
