% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlot.R
\name{rocPlot}
\alias{rocPlot}
\title{Receiver Operating Curve Plot}
\usage{
rocPlot(tpr, fpr)
}
\arguments{
\item{tpr}{True positive rate.}

\item{fpr}{False positive rate.}
}
\value{
A plot.
}
\description{
Plots the precision recall curve.
}
\keyword{binary}
\keyword{curve,}
\keyword{operating}
\keyword{plot,}
\keyword{receiver}
